refactored likes model to work with one view instead of two

+++ profile_app/views.py
@@ -1,6 +1,6 @@
 from django.shortcuts import render, redirect
 from django.contrib import messages
-from .models import Post, Message, Comment
+from .models import Post, Message, Comment, Profile
 from login_app.models import User
 
 
@@ -24,12 +24,14 @@ def update_profile(request, user_id):
             messages.error(request, value)
         return redirect(f'/{user_id}')
     user = User.objects.get(id = user_id)
+    profile = Profile.objects.get(id = user.profile.id)
     user.first_name = request.POST['fname']
     user.last_name = request.POST['lname']
-    user.bio = request.POST['bio']
+    profile.bio = request.POST['bio']
     user.default_img = request.POST['img']
     user.bg_color = request.POST['color']
     user.save()
+    profile.save()
     return redirect(f'/profile/{user_id}')
 
 def follow(request):
@@ -46,7 +48,7 @@ def follow(request):
 
 def create_post(request, user_id):
     if request.method == "GET":
-        return redirect(f'/{user_id}')
+        return redirect('/')
     user = User.objects.get(id= user_id)
     Post.objects.create(
         content = request.POST['post'],
@@ -57,32 +59,45 @@ def create_post(request, user_id):
 def delete_post(request, post_id):
     post_to_delete = Post.objects.get(id = post_id)
     user_id = post_to_delete.added_by.id
+    #post can only be deleted by author
+    if request.session['userid'] != user_id:
+        return redirect('/')
     post_to_delete.delete()
     return redirect(f'/profile/{user_id}')
 
 def like_post(request):
-    #if GET request, redirect
+    #if GET request, redirect 
+    if request.method == "GET":
+        return redirect('/')
+    user = User.objects.get(id = request.session['userid'])
+    post = Post.objects.get(id = request.POST['like'])
     #if user already likes post, unlike post
-    #otherwise, like post
-    pass
+    if user in post.likes.all():
+        post.likes.remove(user)
+    else:
+        post.likes.add(user)
+    return redirect(f'/profile/{post.added_by.id}')
 
 def comment_post(request, post_id):
     post = Post.objects.get(id = post_id)
-    profile = User.objects.get(id = post.added_by.id)
+    if request.method == "GET":
+        return redirect(f'/profile/{post.added_by.id}')
     user = User.objects.get(id = request.session['userid'])
     Comment.objects.create(
         content = request.POST['comment'],
         added_by = user,
         post = post
     )
-    return redirect(f'/profile/{profile.id}')
+    return redirect(f'/profile/{post.added_by.id}')
 
 def delete_post_comment(request, comment_id):
     comment_to_delete = Comment.objects.get(id = comment_id)
+    #comment can only be deleted by author
+    if request.session['userid'] != comment_to_delete.added_by.id:
+        return redirect('/')
     post = Post.objects.get(id = comment_to_delete.post.id)
-    profile = User.objects.get(id = post.added_by.id)
     comment_to_delete.delete()
-    return redirect(f'/profile/{profile.id}')
+    return redirect(f'/profile/{post.added_by.id}')
 
 def like_post_comment(request, comment_id):
     #if GET request, redirect
